requirement.txt
asttokens==2.2.1
blinker==1.6
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
debugpy==1.6.7
executing==1.2.0
Flask==2.3.3
ipykernel==6.25.1
ipython==8.12.0
jedi==0.19.0
llvmlite==0.40.1
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
numba==0.57.1
numpy==1.24.3
pandas==2.0.3
Pillow==9.4.0
psutil==5.9.0
pure-eval==0.2.2
pytz==2023.3
pywin32==305.1
pyzmq==23.2.1
scikit-learn==1.3.0
scipy==1.10.1
shap==0.42.1
slicer==0.0.7
stack-data==0.6.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.0
tqdm==4.66.1